+++ nomad-core/src/aws.rs
@@ -1,13 +1,13 @@
-use once_cell::sync::OnceCell;
 use rusoto_core::{
     credential::{AutoRefreshingProvider, ProvideAwsCredentials},
     Client, HttpClient,
 };
 use rusoto_kms::KmsClient;
 use rusoto_sts::WebIdentityProvider;
+use tokio::sync::OnceCell;
 
-static CLIENT: OnceCell<Client> = OnceCell::new();
-static KMS_CLIENT: OnceCell<KmsClient> = OnceCell::new();
+static CLIENT: OnceCell<Client> = OnceCell::const_new();
+static KMS_CLIENT: OnceCell<KmsClient> = OnceCell::const_new();
 
 // Try to get an irsa provider
 #[tracing::instrument]
@@ -34,37 +34,28 @@ async fn try_irsa_provider() -> Option<AutoRefreshingProvider<WebIdentityProvide
 /// 3. environment
 /// 4. Conf file
 pub async fn get_client() -> &'static Client {
-    // init exactly once
-    if CLIENT.get().is_none() {
-        // try IRSA first
-        let client = match try_irsa_provider().await {
-            Some(credentials_provider) => {
-                let dispatcher = HttpClient::new().unwrap();
-                Client::new_with(credentials_provider, dispatcher)
+    CLIENT
+        .get_or_init(|| async {
+            match try_irsa_provider().await {
+                Some(credentials_provider) => {
+                    let dispatcher = HttpClient::new().unwrap();
+                    Client::new_with(credentials_provider, dispatcher)
+                }
+                // if the IRSA provider returned no creds, use the default
+                // credentials chain
+                None => Client::shared(),
             }
-            // if the IRSA provider returned no creds, use the default
-            // credentials chain
-            None => Client::shared(),
-        };
-
-        if CLIENT.set(client).is_err() {
-            panic!("unable to set Client")
-        };
-    }
-
-    CLIENT.get().expect("just init")
+        })
+        .await
 }
 
 /// Get a shared KMS client
 pub async fn get_kms_client() -> &'static KmsClient {
-    if KMS_CLIENT.get().is_none() {
-        let client = get_client().await.clone();
+    KMS_CLIENT
+        .get_or_init(|| async {
+            let client = get_client().await.clone();
 
-        let kms = KmsClient::new_with_client(client, Default::default());
-
-        if KMS_CLIENT.set(kms).is_err() {
-            panic!("unable to set KmsClient")
-        };
-    }
-    KMS_CLIENT.get().expect("just init")
+            KmsClient::new_with_client(client, Default::default())
+        })
+        .await
 }