+++ tools/killswitch/src/main.rs
@@ -4,21 +4,49 @@ extern crate assert_matches;
 
 mod errors;
 mod killswitch;
-mod output;
+mod secrets;
 mod settings;
 
-use crate::killswitch::KillSwitch;
-use crate::output::Message;
-use crate::{errors::Error, output::Output, settings::Settings};
+use crate::{errors::Error, killswitch::KillSwitch, secrets::Secrets, settings::Settings};
 use clap::{ArgGroup, Parser, ValueEnum};
 use std::{
     env,
     io::{stdout, Write},
     process::exit,
+    sync::mpsc::channel,
 };
+
+/// AWS settings
+const AWS_REGION: &str = "us-west-2";
+const AWS_CREDENTIALS_PROFILE_DEVELOPMENT: &str = "nomad-xyz-dev";
+const AWS_CREDENTIALS_PROFILE_PRODUCTION: &str = "nomad-xyz-prod";
+const CONFIG_S3_BUCKET_DEVELOPMENT: &str = "nomad-killswitch-config-dev";
+const CONFIG_S3_BUCKET_PRODUCTION: &str = "nomad-killswitch-config-prod";
+const CONFIG_S3_KEY: &str = "config.yaml";
+
+/// Local secrets. For testing only
+const SECRETS_PATH_LOCAL_TESTING: &str = concat!(
+    env!("CARGO_MANIFEST_DIR"),
+    "/../../fixtures/killswitch_secrets.testing.yaml"
+);
 
 /// Result returning KillSwitch `Error`
 pub(crate) type Result<T> = std::result::Result<T, Error>;
+
+/// The environment we're targeting
+#[derive(ValueEnum, Clone, Debug, Eq, PartialEq)]
+enum Environment {
+    /// The development environment
+    Development,
+    /// The production environment
+    Production,
+    /// Use local secrets. For testing only
+    #[clap(hide = true)]
+    LocalPath,
+    /// Pull all secrets from environment. For testing only
+    #[clap(hide = true)]
+    AlreadySet,
+}
 
 /// What we're killing, currently only `TokenBridge`
 #[derive(ValueEnum, Clone, Debug)]
@@ -33,11 +61,15 @@ enum App {
     ArgGroup::new("which_networks")
     .required(true)
     .multiple(false)
-    .args(&["all", "all-inbound"])
+    .args(&["all", "all_inbound"])
 ))]
 struct Args {
+    /// Which environment to target
+    #[clap(long, value_enum)]
+    environment: Environment,
+
     /// Which app to kill
-    #[clap(long, arg_enum)]
+    #[clap(long, value_enum)]
     app: App,
 
     /// Kill all available networks
@@ -48,10 +80,11 @@ struct Args {
     #[clap(long, value_name = "NETWORK")]
     all_inbound: Option<String>,
 
-    // The most common form of streaming JSON is line delimited
-    // hide this behind a (hidden) flag so it's not abused
+    /// Actually execute `killswitch`. This is an inverse of
+    /// a `--dry_run` flag and makes more sense given the destructive
+    /// nature of this utility
     #[clap(long, hide = true)]
-    pretty: bool,
+    force: bool,
 }
 
 /// Exit codes as found in <sysexits.h>
@@ -60,56 +93,118 @@ enum ExitCode {
     BadConfig = 78,
 }
 
-/// Print `Output` to stdout as json
-fn report(message: Message, pretty: bool) {
-    let command = env::args().collect::<Vec<_>>().join(" ");
-    let output = Output { command, message };
-    let json = if pretty {
-        serde_json::to_string_pretty(&output)
-    } else {
-        serde_json::to_string(&output)
-    }
-    .expect("Serialization error. Should never happen");
-    stdout()
-        .lock()
-        .write_all(&format!("{}\n", json).into_bytes())
-        .expect("Write to stdout error. Should never happen");
+fn write_stdout(message: &str) {
+    stdout().lock().write_all(message.as_bytes()).unwrap()
 }
 
 /// KillSwitch entry point
-#[tokio::main(flavor = "current_thread")]
-async fn main() {
+#[tokio::main(flavor = "multi_thread")]
+async fn main() -> Result<()> {
     let args = Args::parse();
-    let pretty = args.pretty;
+    let command = env::args().collect::<Vec<_>>().join(" ");
+    write_stdout("\n");
+    write_stdout(&format!("Running `{}`\n", command));
 
-    // Try to build `NomadConfig`, exiting immediately if we can't
+    if Environment::AlreadySet != args.environment {
+        write_stdout("Fetching secrets from S3 using local AWS credentials... ");
+        let secrets = match &args.environment {
+            Environment::Development => {
+                Secrets::fetch(
+                    AWS_CREDENTIALS_PROFILE_DEVELOPMENT,
+                    AWS_REGION,
+                    CONFIG_S3_BUCKET_DEVELOPMENT,
+                    CONFIG_S3_KEY,
+                )
+                .await
+            }
+            Environment::Production => {
+                Secrets::fetch(
+                    AWS_CREDENTIALS_PROFILE_PRODUCTION,
+                    AWS_REGION,
+                    CONFIG_S3_BUCKET_PRODUCTION,
+                    CONFIG_S3_KEY,
+                )
+                .await
+            }
+            Environment::LocalPath => Secrets::load(SECRETS_PATH_LOCAL_TESTING).await,
+            Environment::AlreadySet => unreachable!(),
+        };
+        if let Err(error) = secrets {
+            write_stdout(&format!("Failed: {}\n", error));
+            exit(ExitCode::BadConfig as i32)
+        }
+        write_stdout("Ok\n");
+
+        // Set `Secrets` as environment variables for `Settings` to pick up
+        secrets.unwrap().set_environment();
+    }
+
+    write_stdout("Building settings from environment... ");
     let settings = Settings::new().await;
     if let Err(error) = settings {
-        // We've hit a blocking error, bail
-        report(error.into(), pretty);
+        write_stdout(&format!("Failed: {}\n", error));
         exit(ExitCode::BadConfig as i32)
     }
+    write_stdout("Ok\n");
 
-    // Try to build `KillSwitch`. If we hit `NoNetworks` error, nothing to do, bail
-    let killswitch = KillSwitch::new(args, settings.unwrap()).await;
+    let settings = settings.unwrap();
+
+    write_stdout("Checking `killswitch` for killable networks... ");
+    let killswitch = KillSwitch::new(&args, settings).await;
     if let Err(error) = killswitch {
-        // We've hit a blocking error, bail
-        report(error.into(), pretty);
+        write_stdout(&format!("Failed: {}\n", error));
         exit(ExitCode::BadConfig as i32)
     }
+    write_stdout("Ok\n");
 
-    // Get errors that block individual channels, report before proceeding. Do not bail
-    let (killswitch, errors) = killswitch.unwrap().get_blocking_errors().await;
-    if let Some(errors) = errors {
-        // Stream these blocking errors before running transactions
-        // so users can be updated as fast as possible
-        report(errors, pretty);
+    let killswitch = killswitch.unwrap();
+
+    write_stdout("\n");
+    write_stdout("`killswitch` is ready to attempt to kill the selected channels:\n");
+    for channel in killswitch.channels() {
+        write_stdout(&format!(
+            "[CHANNEL] {} -> {}\n",
+            channel.home, channel.replica
+        ));
     }
 
-    // Run all channels that *could* succeed
-    let results = killswitch.run().await;
-
-    // Give users final results, exit ok
-    report(results, pretty);
-    exit(ExitCode::Ok as i32)
+    if !&args.force {
+        write_stdout("\n");
+        write_stdout("[NOTICE] Nothing killed!\n");
+        write_stdout("\n");
+        write_stdout("To kill the selected networks, run the same command again with the `--force` flag added.\n");
+        write_stdout("\n");
+        exit(ExitCode::Ok as i32)
+    } else {
+        write_stdout("\n");
+        write_stdout("Running `killswitch`...\n");
+        let (tx, rx) = channel();
+        let handles = killswitch.run(tx);
+        for _ in 0..killswitch.channel_count() {
+            let (channel, result) = rx.recv().unwrap();
+            write_stdout("\n");
+            write_stdout(&format!(
+                "[CHANNEL] {} -> {}\n",
+                channel.home, channel.replica
+            ));
+            match result {
+                Ok(txid) => {
+                    write_stdout(&format!(
+                        "[SUCCESS] transaction id for unenrollment: {:?}\n",
+                        txid
+                    ));
+                }
+                Err(error) => {
+                    write_stdout(&format!("[FAILURE] {}\n", error));
+                }
+            }
+        }
+        write_stdout("\n");
+        for handle in handles {
+            handle
+                .await
+                .expect("Should not happen. Errors should have been caught");
+        }
+        exit(ExitCode::Ok as i32)
+    }
 }