+++ tools/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+- adds the ability for killswitch to auto-configure
+- makes killswitch output human readable
+- makes killswitch execute transactions in parallel to improve speed
 - adds a killswitch binary that allows for the manual shutdown of bridge channels 
 - adds tests and test fixtures for killswitch 
 - adds a readme, including documentation for killswitch